mysite/requirements.txt
Django
django-solo
docker
psycopg2-binary
pillow
gunicorn
djangorestframework
markdown
django-filter
